Revert "Refactor loader tasks into modular subpackages (#84)"
This reverts commit 9d786a1745dcbb76fc7400d7faee50029f0fdb21.
