+++ tests/test_loader_logging.py
@@ -34,7 +34,7 @@ def test_run_logs_upsert(monkeypatch, caplog):
     with caplog.at_level(logging.INFO, logger="mcp_plex.loader"):
         asyncio.run(loader.run(None, None, None, sample_dir, None, None))
     assert "Loaded 2 items" in caplog.text
-    assert "Storage worker" in caplog.text
+    assert "Upsert worker" in caplog.text
     assert "handling 2 points" in caplog.text
     assert "processed 2 items" in caplog.text
